move BooleanCell to generic components directory
