+++ frontend/src/components/BooleanCell.tsx
